Implemented saving forecast data from OpenWeatherMap to DB
Implemented:
- saving forecast data from OpenWeatherMap to DB
- OWMForecastPresenter
- OWMForecastInteractor
- OWMForecastMapper
- OWMForecastRepository
- OWMForecastRealmService
- OWM models for saving to DB
Changed:
- versionName to 0.25
- added and changed OWM typealiases
- other minor changes

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/currentWeather/IOWMCurrentWeatherResponse.kt
@@ -25,7 +25,7 @@ interface IOWMCurrentWeatherResponse : ICurrentWeatherResponse {
     val clouds: IOWMCloudsResponse?
 
     /**Time of data calculation, unix, UTC*/
-    val timeOfDataCalculationUnixUTC: Int
+    val timeOfDataCalculationUnixUTC: Long
 
     val sys: OWMSysResponse?
