+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/currentWeather/IOWMSysResponse.kt
@@ -17,9 +17,9 @@ interface IOWMSysResponse {
     val countryCode: String
 
     /**Sunrise time, unix, UTC*/
-    val sunrise: Int
+    val sunrise: Long
 
     /**Sunset time, unix, UTC*/
-    val sunset: Int
+    val sunset: Long
 
 }
