+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/forecast/IOWMCityResponse.kt
@@ -13,6 +13,6 @@ interface IOWMCityResponse {
     val name: String
 
     /**City ID*/
-    val id: Double
+    val id: Int
 
 }
