+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/yandexWeather/IYWHourInfoResponse.kt
@@ -6,7 +6,7 @@ interface IYWHourInfoResponse {
     val hourInLocalTime: String
 
     /**The time of the forecast in Unix time.*/
-    val hourInUnixTime: Int
+    val hourInUnixTime: Long
 
     /**Temperature (°C).*/
     val temperature: Double
