+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/implementation/openWeatherMap/forecast/OWMCityResponse.kt
@@ -22,6 +22,6 @@ data class OWMCityResponse(
 
     @SerializedName("id")
     @Expose
-    override val id: Double
+    override val id: Int
 
 ) : IOWMCityResponse
